requirements/staging.txt
# Production requirements first
-r production.txt

# Essential Performance Tools
# ------------------------------------------------------------------------------
django-extensions==3.2.3  # Development utilities, shell_plus, runserver_plus

# Optional Performance Tools (uncomment as needed)
# ------------------------------------------------------------------------------
# django-silk==5.1.0  # Advanced profiling and inspection of live requests
# django-cachalot==3.1.0  # Cache analysis and optimization
# django-query-profiler==0.7.0  # Detailed query analysis

# Essential Testing
# ------------------------------------------------------------------------------
pytest==8.3.4  # Test framework
pytest-django==4.9.0  # Django test utilities
coverage==7.6.10  # Code coverage analysis

# Optional Testing (uncomment as needed)
# ------------------------------------------------------------------------------
# pytest-benchmark==4.0.0  # Performance benchmarking
# pytest-xdist==3.5.0  # Parallel test execution
# pytest-cov==4.1.0  # Coverage reporting
# django-coverage-plugin==3.1.0  # Django-specific coverage

# Essential Development
# ------------------------------------------------------------------------------
ipdb==0.13.13  # Enhanced debugging

# Optional Development (uncomment as needed)
# ------------------------------------------------------------------------------
# Werkzeug[watchdog]==3.1.3  # Development server with auto-reload
# memory-profiler==0.61.0  # Memory usage analysis
# line-profiler==4.1.1  # Line-by-line performance analysis
